refactor: fix code quality issues and integrate inbound queue into daemon
- Extract fallback branch from enqueue_inbound into process_inbound_direct
  (enqueue_inbound body: 69→26 lines, within 50-line limit)
- Split impl SlashResult (122→53 lines) by moving execute_system_append,
  execute_new_session, execute_stop to side_effect.rs as free functions
- Extract busy reply text as const BUSY_REPLY_TEXT
- Remove #[allow(dead_code)] from inbound_tx field
- Call gateway.start_inbound_queue() in daemon init_phase_3_core_services
  to wire inbound queue into the startup flow

Source: user
Type: refactor
